requirements.txt
streamlit
pandas
numpy
yfinance
matplotlib
plotly
openpyxl
scipy
streamlit_calendar
bcrypt == 3.1.7
captcha == 0.5.0
cryptography == 42.0.5
PyJWT == 2.3.0
PyYAML == 5.3.1
extra-streamlit-components == 0.1.70
streamlit-authenticator == 0.4.2
requests
curl-cffi>=0.5.10

# === AJOUTS POUR GOOGLE SHEETS ===
gspread == 5.12.0
google-auth == 2.23.4
google-auth-oauthlib == 1.1.0
google-auth-httplib2 == 0.2.0

# === OPTIONNEL (pour fonctionnalités futures) ===
# stripe == 7.8.0
# sendgrid == 6.10.0
# redis == 5.0.1
